Extend amc with conditions, codecs and a block hash, and rebuild AMS
amc:

- fcond: list membership tied to a condition on a field.  One row declares that
  while a condition holds, a record belongs on a list, and the field's generated
  setter performs the move on every transition.  Written by hand the insert and
  the remove live in two places and nothing pairs them; the row is the pairing.
- ffunc: replaces fcleanup and fuserinit.  Its key is a field and a tfunc name,
  so every generated function a field can carry is selected through one table
  instead of one table per kind of function.
- ccmdline: replaces fcmdline.  A command ctype names the base cmdline it
  extends, so one tool inherits another's options as a relation between ctypes.
- Blkhash: a cross-reference index for a key with a dense integer component.
  The low bits of that component address a block and the rest of the key is
  hashed, so an index of ten million consecutive sequence numbers costs one
  pointer per block rather than per key, and neighbours share a cache line.
- msg, msgfield and payloadhdr: amc derives each message ctype's wire layout --
  its length, and the offset, width and endianness of every leaf field -- and
  writes it back as data.  The frame is then readable from the schema instead of
  from the generated code.
- pbuf: a protobuf codec generated from ssim.  cpbuf, fpbuf, fpbufcase,
  pbsyntax and amcdb.pbtype describe how a ctype maps onto protobuf wire types,
  and amc emits a self-contained encoder and decoder -- no protoc, no upstream
  runtime.
- cjson, cjsbltin and jsbltin: JSON interfaces and serde helpers for the
  TypeScript side of a namespace tagged nsjs, alongside its ctypes and enums.
- amcdb.cbtype names the lifecycle events a user function may handle on a field;
  amcdb.tcond names the conditions a tclass reports, which is the vocabulary
  fcond binds an index to.

acr_ed:

- -create -fcurs adds a cursor to a field, and -create -dispatch_msg adds one
  message to a dispatch, so neither has to be inserted by hand.
- -create -fstep takes -steptype, and a new source or script file takes
  -license.
- -amc, -pooltype and -test are gone: amc runs as part of every edit that needs
  it, a pool's type follows from the reftype, and a test is created through the
  test tables.

AMS:

- lib_ams is rewritten around the group id.  A group is a shared-memory ring
  between two processes named by proctype-nodeidx-procidx.grptype-grpidx, and
  amsdb.grptype is the registry of ring kinds -- a log ring, a supervisor ring,
  a ring from a parent process to the user process it launched.  amsdb.shmtype
  is gone; a shm is a group now.
- A message board carries payloads too large for a ring.  The writer places one
  copy on its board and posts a reference in each reader's ring, so a message
  reaching N readers is written once, and the slot is freed when the last reader
  has passed it.
- A process may bridge to a child process, reading and writing the child's rings
  under its own identity, and may take input from a file descriptor as ssim text
  when no ring is available -- which is what lets a filter run at the end of a
  pipe.
- A message a full ring cannot take is queued on that ring and written by
  lib_ams's own step as budget appears, so a caller has neither a drop to handle
  nor a block to wait through.
- amsspy reads the shared-memory segments on a host: what exists, who writes
  each one, how far behind each reader is, and -clean reclaims the segments
  whose writer is gone.
- samp_meng, the sample matching engine, speaks the AMS protocol: its messages
  are ams.SampMeng%Msg and it reads and writes them through lib_ams, as binary
  on a ring or as ssim text on a pipe.

Documentation:

- amc's internals.md, 6700 lines of it, is one page per feature: cfmt, cmdline,
  dispatch, fbuf, fcond, hook, inheritance, io, js, pbuf, reflection, trace,
  varlen and the rest, each with its schema rows and a worked example.
- reftypes.md is a directory with a page per reftype, from Alias to ZSListMT,
  each stating what the field stores, which functions amc emits for it, and what
  it costs.
- The backend is documented separately: the generation pipeline, the tclass and
  tfunc model, the data model it walks, and how to extend it.
- acr's and acr_ed's internals pages fold into their READMEs, and txt/openacr is
  a new guide set -- setup, ssim, acr, schema, codegen, runtime, build and
  recipes -- each written for the task rather than for the tool.

Coverage and packaging:

- dev.uncovfunc records every function the suite leaves entirely unhit, so the
  untested surface is a query rather than a report to read.
- dev.pkggen records what a package generates, and apm -generate writes a
  package's records to apm/gen/<package>.ssim.

+++ cpp/lib/algo/line.cpp
@@ -54,7 +54,7 @@ bool algo::LinebufNext(LineBuf &L, strptr &result) {
     // beg,end indicate location of eol inside the incoming string
     // search for \n, or to end of incoming string
     char *incoming=L.incoming.elems;
-    int beg=0,end=L.incoming.n_elems;
+    i64 beg=0,end=L.incoming.n_elems;
     for (; beg<end; beg++) {
         if (L.incoming.elems[beg]=='\n') {
             end=beg+1;
@@ -82,7 +82,7 @@ bool algo::LinebufNext(LineBuf &L, strptr &result) {
 // (which is either \n or \r\n).
 strptr algo::GetLine(StringIter &S) {
     strptr ret(S.Rest());
-    for (int i=0; i<ret.n_elems; i++) {
+    for (i64 i=0; i<ret.n_elems; i++) {
         S.index++;// skip character in S
         if (ret[i]=='\n') {
             ret.n_elems = i-(i>0 && ret[i-1]=='\r');
